IDB: validate a parameter in Connection::GetAll mojom
A DCHECK is converted to a CHECK, but should no longer be reachable
even with a compromised renderer.

Bug: 459129408
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7143018
Commit-Queue: Evan Stade <evanstade@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1548219}

+++ content/browser/indexed_db/instance/connection.h
@@ -167,7 +167,7 @@ class CONTENT_EXPORT Connection : public blink::mojom::IDBDatabase {
               int64_t index_id,
               blink::IndexedDBKeyRange key_range,
               blink::mojom::IDBGetAllResultType result_type,
-              int64_t max_count,
+              uint32_t max_count,
               blink::mojom::IDBCursorDirection direction,
               blink::mojom::IDBDatabase::GetAllCallback callback) override;
   void OpenCursor(
