+++ content/browser/indexed_db/instance/database.h
@@ -180,7 +180,7 @@ class CONTENT_EXPORT Database {
       int64_t index_id,
       blink::IndexedDBKeyRange key_range,
       blink::mojom::IDBGetAllResultType result_type,
-      int64_t max_count,
+      uint32_t max_count,
       blink::mojom::IDBCursorDirection direction,
       blink::mojom::IDBDatabase::GetAllCallback callback,
       Transaction* transaction);
@@ -259,7 +259,7 @@ class CONTENT_EXPORT Database {
                          int64_t index_id,
                          blink::IndexedDBKeyRange key_range,
                          blink::mojom::IDBGetAllResultType result_type,
-                         int64_t max_count,
+                         uint32_t max_count,
                          blink::mojom::IDBCursorDirection direction,
                          std::unique_ptr<GetAllResultSinkWrapper> result_sink,
                          Transaction* transaction);
